Can configure sim using GUI, fixed race cond. on entry space

+++ src/main/java/aim4/map/cpm/CPMBasicMap.java
@@ -89,7 +89,7 @@ public abstract class CPMBasicMap implements CPMMap{
     /**The set of intersections */
     protected List<SimpleIntersection> intersections = new ArrayList<SimpleIntersection>();
 
-    public CPMBasicMap(int laneWidth, double speedLimit,
+    public CPMBasicMap(double laneWidth, double speedLimit,
                        double initTime) {
 
         this.laneWidth = laneWidth;
